fix: ajusta tipo do score

+++ scripts/migrate.ts
@@ -18,7 +18,7 @@ import { run } from '../src/db';
       id SERIAL PRIMARY KEY,
       user_id INTEGER NOT NULL,
       time_ms INTEGER NOT NULL,
-      score INTEGER NOT NULL,
+      score DECIMAL(15,4) NOT NULL,
       created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
       FOREIGN KEY (user_id) REFERENCES users(id)
     )
